frontend/requirements.txt
altair==4.2.0; python_version >= "3.7"
appnope==0.1.2; platform_system == "Darwin" and python_version >= "3.8" and sys_platform == "darwin"
argon2-cffi-bindings==21.2.0; python_version >= "3.7"
argon2-cffi==21.3.0; python_version >= "3.7"
asttokens==2.0.5; python_version >= "3.8"
attrs==21.4.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
backcall==0.2.0; python_version >= "3.8"
backports.zoneinfo==0.2.1; python_version < "3.9" and python_version >= "3.6" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
beautifulsoup4==4.10.0; python_full_version > "3.0.0" and python_version >= "3.7"
bleach==4.1.0; python_version >= "3.7"
blinker==1.4; python_version >= "3.6"
cachetools==5.0.0; python_version >= "3.7" and python_version < "4.0"
certifi==2021.10.8; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
cffi==1.15.0; python_full_version >= "3.6.1" and python_version >= "3.7" and implementation_name == "pypy"
charset-normalizer==2.0.12; python_full_version >= "3.6.0" and python_version >= "3.6"
click==8.0.4; python_version >= "3.6"
colorama==0.4.4; python_version >= "3.8" and python_full_version < "3.0.0" and platform_system == "Windows" and sys_platform == "win32" or platform_system == "Windows" and python_version >= "3.8" and python_full_version >= "3.5.0" and sys_platform == "win32"
debugpy==1.6.0; python_version >= "3.7"
decorator==5.1.1; python_version >= "3.8"
defusedxml==0.7.1; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
entrypoints==0.4; python_full_version >= "3.6.1" and python_version >= "3.7"
executing==0.8.3; python_version >= "3.8"
gitdb==4.0.9; python_version >= "3.7"
gitpython==3.1.27; python_version >= "3.7"
idna==3.3; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
importlib-metadata==4.11.3; python_version >= "3.7"
importlib-resources==5.6.0; python_version < "3.9" and python_version >= "3.7"
ipykernel==6.10.0; python_version >= "3.7"
ipython-genutils==0.2.0; python_version >= "3.7"
ipython==8.2.0; python_version >= "3.8"
ipywidgets==7.7.0; python_version >= "3.7"
jedi==0.18.1; python_version >= "3.8"
jinja2==3.1.1; python_version >= "3.7"
jsonschema==4.4.0; python_version >= "3.7"
jupyter-client==7.1.2; python_full_version >= "3.7.0" and python_version >= "3.7"
jupyter-core==4.9.2; python_full_version >= "3.6.1" and python_version >= "3.7"
jupyterlab-pygments==0.1.2; python_version >= "3.7"
jupyterlab-widgets==1.1.0; python_version >= "3.7"
markupsafe==2.1.1; python_version >= "3.7"
matplotlib-inline==0.1.3; python_version >= "3.8"
mistune==0.8.4; python_version >= "3.7"
nbclient==0.5.13; python_full_version >= "3.7.0" and python_version >= "3.7"
nbconvert==6.4.5; python_version >= "3.7"
nbformat==5.2.0; python_full_version >= "3.7.0" and python_version >= "3.7"
nest-asyncio==1.5.4; python_full_version >= "3.7.0" and python_version >= "3.7"
notebook==6.4.10; python_version >= "3.7"
numpy==1.22.3
packaging==21.3; python_version >= "3.7"
pandas==1.4.1; python_version >= "3.8"
pandocfilters==1.5.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.7"
parso==0.8.3; python_version >= "3.8"
pexpect==4.8.0; sys_platform != "win32" and python_version >= "3.8"
pickleshare==0.7.5; python_version >= "3.8"
pillow==9.0.1; python_version >= "3.7"
prometheus-client==0.13.1; python_version >= "3.7"
prompt-toolkit==3.0.28; python_full_version >= "3.6.2" and python_version >= "3.8"
protobuf==3.19.4; python_version >= "3.6"
psutil==5.9.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.7"
ptyprocess==0.7.0; sys_platform != "win32" and python_version >= "3.8" and os_name != "nt"
pure-eval==0.2.2; python_version >= "3.8"
py==1.11.0; python_full_version >= "3.6.1" and python_version >= "3.7" and implementation_name == "pypy"
pyarrow==7.0.0; python_version >= "3.7"
pycparser==2.21; python_full_version >= "3.6.1" and python_version >= "3.7" and implementation_name == "pypy"
pydeck==0.7.1; python_version >= "3.7"
pygments==2.11.2; python_version >= "3.8"
pympler==1.0.1; python_version >= "3.6"
pyparsing==3.0.7; python_version >= "3.6"
pyrsistent==0.18.1; python_version >= "3.7"
python-dateutil==2.8.2; python_full_version >= "3.6.1" and python_version >= "3.8"
pytz-deprecation-shim==0.1.0.post0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
pytz==2022.1; python_version >= "3.8"
pywin32==303; sys_platform == "win32" and platform_python_implementation != "PyPy" and python_full_version >= "3.6.1" and python_version >= "3.7"
pywinpty==2.0.5; os_name == "nt" and python_version >= "3.7"
pyzmq==22.3.0; python_full_version >= "3.6.1" and python_version >= "3.7"
requests==2.27.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
semver==2.13.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
send2trash==1.8.0; python_version >= "3.7"
six==1.16.0; python_version >= "3.8" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.8"
smmap==5.0.0; python_version >= "3.7"
soupsieve==2.3.1; python_full_version > "3.0.0" and python_version >= "3.7"
stack-data==0.2.0; python_version >= "3.8"
streamlit==1.8.0; python_version >= "3.6"
terminado==0.13.3; python_version >= "3.7"
testpath==0.6.0; python_version >= "3.7"
toml==0.10.2; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
toolz==0.11.2; python_version >= "3.7"
tornado==6.1; python_full_version >= "3.6.1" and python_version >= "3.7"
traitlets==5.1.1; python_full_version >= "3.7.0" and python_version >= "3.8"
tzdata==2022.1; platform_system == "Windows" and python_version >= "3.6" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
tzlocal==4.1; python_version >= "3.6"
urllib3==1.26.9; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.6"
validators==0.18.2; python_version >= "3.6"
watchdog==2.1.7; python_version >= "3.6"
wcwidth==0.2.5; python_full_version >= "3.6.2" and python_version >= "3.8"
webencodings==0.5.1; python_version >= "3.7"
widgetsnbextension==3.6.0; python_version >= "3.7"
zipp==3.7.0; python_version < "3.9" and python_version >= "3.7"
